requirements.txt
# Flask & Web
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
orjson==3.11.3

# Google Cloud
google-cloud-bigquery
google-cloud-aiplatform>=1.66.0
google-auth==2.41.1

# Data processing
pyarrow>=16.1.0
db-dtypes==1.4.3
pandas>=2.1.4
openpyxl>=3.1.0
numpy>=1.26.2
scikit-learn>=1.3.2

# LDAP Authentication
ldap3>=2.9.1
